created post category Api and established relation between product and cayegory

+++ src/main/java/com/example/EcommerceSpring/gateway/api/FakeStoreProductApi.java
@@ -17,7 +17,7 @@ public interface FakeStoreProductApi {
     Call<FakeStoreProductResponseDTO> getProductsByCategory(@Query("type") String type) throws IOException;
 
     @GET("products/{id}")
-    Call<FakeStoreSingleProductResponseDTO> getProduct(@Path("id") int id) throws IOException;
+    Call<FakeStoreSingleProductResponseDTO> getProduct(@Path("id") Long id) throws IOException;
 
 
 }
